# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
...
    "/backups": {
      "get": {
        "operationId": "BackupsController_list",
        "parameters": [
          {
            "name": "deploymentId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BackupDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "backups"
        ]
      },
      "post": {
        "operationId": "BackupsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBackupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "backups"
        ]
      }
    },
    "/backups/volumes": {
      "get": {
        "operationId": "BackupsController_volumes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolumesDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "backups"
        ]
      }
    },
    "/backups/{id}/restore": {
      "post": {
        "operationId": "BackupsController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "backups"
        ]
      }
    },
    "/backups/{id}/push/{destinationId}": {
      "post": {
        "operationId": "BackupsController_push",
        "parameters": [
          {
            "name": "destinationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "backups"
        ]
      }
    },
    "/backups/{id}/download": {
      "get": {
        "operationId": "BackupsController_download",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "backups"
        ]
      }
    },
    "/backups/{id}": {
      "delete": {
        "operationId": "BackupsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "backups"
        ]
      }
    },
...
    "/admin/images": {
      "get": {
        "operationId": "AdminController_getImages",
        "parameters": [
          {
            "name": "all",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminImageDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/images/{id}": {
      "delete": {
        "operationId": "AdminController_deleteImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Image ID or tag.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Image removed."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/images/prune": {
      "post": {
        "operationId": "AdminController_pruneImages",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PruneResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/containers": {
      "get": {
        "operationId": "AdminController_getContainers",
        "parameters": [
          {
            "name": "all",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminContainerDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/containers/prune": {
      "post": {
        "operationId": "AdminController_pruneContainers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PruneResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/settings": {
      "get": {
        "operationId": "AdminController_getSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettingsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "put": {
        "operationId": "AdminController_updateSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettingsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Willy API",
    "description": "",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OkResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "ok"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "SessionUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "OPERATOR",
              "VIEWER"
            ]
          }
        },
        "required": [
          "id",
          "email",
          "role"
        ]
      },
      "SessionDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/SessionUserDto"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "user"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "OPERATOR",
              "VIEWER"
            ]
          }
        },
        "required": [
          "userId",
          "email",
          "role"
        ]
      },
...
      "BackupDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "deploymentId": {
            "type": "string",
            "nullable": true
          },
          "kind": {
            "type": "string",
            "enum": [
              "VOLUME_TAR",
              "PG_DUMP",
              "S3_SYNC"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "RUNNING",
              "SUCCESS",
              "FAILED"
            ]
          },
          "target": {
            "type": "string",
            "nullable": true
          },
          "sizeBytes": {
            "type": "number",
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "offsiteUrl": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "deploymentId",
          "kind",
          "status",
          "target",
          "sizeBytes",
          "errorMessage",
          "offsiteUrl",
          "createdAt",
          "finishedAt"
        ]
      },
      "VolumesDto": {
        "type": "object",
        "properties": {
          "volumes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "volumes"
        ]
      },
      "CreateBackupDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "VOLUME_TAR",
              "PG_DUMP",
              "S3_SYNC"
            ]
          },
          "target": {
            "type": "string",
            "example": "willy_compose-blog_data"
          },
          "deploymentId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "kind",
          "target"
        ]
      },
...
      "DeploymentRefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "AdminImageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Full Docker image ID (sha256:…)."
          },
          "repoTags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "size": {
            "type": "number",
            "description": "Actual on-disk size of the image in bytes."
          },
          "virtualSize": {
            "type": "number",
            "description": "Virtual size including shared layers, in bytes."
          },
          "created": {
            "type": "number",
            "description": "Unix timestamp when the image was created."
          },
          "deployments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeploymentRefDto"
            }
          },
          "activeContainersCount": {
            "type": "number",
            "description": "Number of containers based on this image."
          },
          "managed": {
            "type": "boolean",
            "description": "Built by Willy or run by a managed deployment."
          }
        },
        "required": [
          "id",
          "repoTags",
          "size",
          "virtualSize",
          "created",
          "deployments",
          "activeContainersCount",
          "managed"
        ]
      },
      "PruneResultDto": {
        "type": "object",
        "properties": {
          "spaceReclaimedBytes": {
            "type": "number",
            "description": "Bytes reclaimed by the prune operation."
          },
          "itemsRemoved": {
            "type": "number",
            "description": "Number of images or containers removed."
          }
        },
        "required": [
          "spaceReclaimedBytes",
          "itemsRemoved"
        ]
      },
      "AdminContainerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "image": {
            "type": "string",
            "description": "Image tag or ID the container was started from."
          },
          "state": {
            "type": "string",
            "description": "Docker state: running, exited, created, paused, …"
          },
          "status": {
            "type": "string",
            "description": "Human-readable status, e.g. \"Up 2 hours\"."
          },
          "created": {
            "type": "number",
            "description": "Unix timestamp when the container was created."
          },
          "deployment": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DeploymentRefDto"
              }
            ]
          },
          "managed": {
            "type": "boolean",
            "description": "Belongs to a Willy-managed deployment."
          }
        },
        "required": [
          "id",
          "name",
          "image",
          "state",
          "status",
          "created",
          "managed"
        ]
      },
      "AppSettingsDto": {
        "type": "object",
        "properties": {
          "showAllResources": {
            "type": "boolean",
            "description": "Show all host containers/images on the Images/Containers pages, not only Willy-managed ones."
          }
        },
        "required": [
          "showAllResources"
        ]
      },
      "UpdateAppSettingsDto": {
        "type": "object",
        "properties": {
          "showAllResources": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}